Pull out reusable components from main Chat component.
