+++ packages/dialect-react-ui/components/ChatButton/index.tsx
@@ -53,7 +53,7 @@ function useOutsideAlerter(
       // Unbind the event listener on clean up
       document.removeEventListener('mousedown', handleClickOutside);
     };
-  }, [ref]);
+  }, [ref, bellRef, setOpen]);
 }
 
 function WrappedChatButton(
